refactor(tests): split test_codegraph_query_tool.py (1980→6 files, phase8 slice)
Split 3 oversized classes across 6 focused modules (all under 500 lines),
with a shared helper module for plain test-utility functions:

- _codegraph_query_helpers.py — _make_def, _patch_resolver_with helpers (52 lines)
- test_codegraph_query_dsl.py — TestParseChain: DSL parser/argument helpers (178 lines)
- test_codegraph_query_tool_core.py — schema, validation, execute basics (395 lines)
- test_codegraph_query_tool_advanced.py — filter/fallback, batch, sort, answer (414 lines)
- test_codegraph_query_tool_has.py — has-step, relation cache, apply_step guard (180 lines)
- test_codegraph_query_internals.py — filter helpers, query state, resolve_query (406 lines)
- test_codegraph_query_internals_facets.py — relation steps, sort, file entries (428 lines)

79 tests pass before and after; full unit suite (16 598 tests) green.
